delete medias in chat

+++ src/common/utils/utils.deleteFile.ts
@@ -3,12 +3,26 @@ import fs from 'fs'
 import { publicDir } from '../../app'
 
 const deleteFile = (filename: string) => {
-    const filePath = path.join(publicDir, filename)
+  const filePath = path.join(publicDir, filename)
+
+  // Check if file exists
+  fs.access(filePath, fs.constants.F_OK, accessErr => {
+    if (accessErr) {
+      if (accessErr.code === 'ENOENT') {
+        console.warn(`File not found: ${filePath}`)
+      } else {
+        console.error('Error checking file accessibility', accessErr)
+      }
+      return
+    }
+
+    // If file exists, delete it
     fs.unlink(filePath, unlinkErr => {
-        if (unlinkErr) {
-            console.error('File couldnt be deleted', unlinkErr)
-        }
+      if (unlinkErr) {
+        console.error('Failed to delete file', unlinkErr)
+      }
     })
+  })
 }
 
 export default deleteFile
